refactor: rename dashboard command module to tui
